refactor: change architecture app
