+++ app/components/user/domain/interfaces/response-reg.interface.ts
