+++ app/components/user/domain/interfaces/user.interface.ts
