+++ app/libraries/common/decoded-message.ts
@@ -1,6 +1,6 @@
 import WebSocket from 'ws';
-import { jsonConverter } from '../utils/jsonConverter';
-import { StructureMessage } from '../models/interfaces/structureMessage.interface';
+import { jsonConverter } from '../utils/json-converter';
+import { StructureMessage } from '../models/interfaces/structure-message.interface';
 
 export const decodedMessage = (message: WebSocket.RawData) => {
   const messageUtf8 = message.toString();
