+++ app/libraries/models/interfaces/structure-message.interface.ts
