+++ app/libraries/models/types/keys-interfaces.ts
@@ -1,0 +1,3 @@
+import { TypesMessages } from '../../../components/user/domain/interfaces/types-messages.interface';
+
+export type KeysOfTypesMessages = keyof TypesMessages;
