+++ app/libraries/router/router.ts
@@ -1,6 +1,6 @@
 import { WebSocket } from 'ws';
-import { typeOfMessage } from '../common/typeOfMessage';
-import { KeysOfTypesMessages } from '../models/types/keysOfInterfaces';
+import { typeOfMessage } from '../common/type-message';
+import { KeysOfTypesMessages } from '../models/types/keys-interfaces';
 
 export const router = (type: string, payload: string, ws: WebSocket) => {
   const handlerMessage = typeOfMessage[type as KeysOfTypesMessages];
