+++ app/libraries/utils/create-send-message.ts
@@ -1,4 +1,4 @@
-import { StructureMessage } from '../models/interfaces/structureMessage.interface';
+import { StructureMessage } from '../models/interfaces/structure-message.interface';
 
 export const createSendMessage = (type: string, data: string): StructureMessage => (
   { type, data, id: 0 });
