+++ nodemon.json
@@ -2,5 +2,5 @@
     "watch": ["src"],
     "ext": ".ts",
     "ignore": [],
-    "exec": "npx ts-node ./src/index.ts"
+    "exec": "npx ts-node ./app/index.ts"
 }
