+++ package.json
@@ -17,9 +17,10 @@
     "@typescript-eslint/eslint-plugin": "^5.60.1",
     "@typescript-eslint/parser": "^5.60.1",
     "cross-env": "^7.0.3",
-    "eslint": "^8.43.0",
+    "eslint": "^8.44.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-plugin-unicorn": "^47.0.0",
     "nodemon": "^2.0.22",
     "ts-loader": "^9.4.4",
     "ts-node": "^10.9.1",
