--- src/models/types/keysOfInterfaces.ts
@@ -1,3 +1,0 @@
-import { TypesMessages } from '../interfaces/typesMessages.interface';
-
-export type KeysOfTypesMessages = keyof TypesMessages;
