+++ tsconfig.json
@@ -4,7 +4,7 @@
         "moduleResolution": "node",
         "target": "ES2020",
         "sourceMap": true,
-        "rootDir": "./src",
+        "rootDir": "./app",
         "outDir": "./dist/",
         "esModuleInterop": true,
         "strict": true,
@@ -12,7 +12,7 @@
         "removeComments": true,
         "allowJs": true
     },
-    "include": ["src/**/*"],
+    "include": ["app/**/*"],
     "exclude": ["node_modules"],
     "ts-node": {
         "esm": true
