+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-  entry: './src/index.ts',
+  entry: './app/index.ts',
   target: 'node',
   output: {
     filename: 'bundle.js',
